requirements.txt
Pillow
requests
exchangelib
pytz
pyyaml
feedparser
dateutil
